Tools/requirements.txt
alabaster==0.7.13
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.0.5
async-lru==2.0.2
attrs==21.4.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.11.1
bibtexparser==1.4.0
bleach==5.0.1
canon-tools==0.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
chemprop==1.5.2
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.5
cycler==0.11.0
debugpy==1.6.1
decorator==5.1.1
defusedxml==0.7.1
docstring-parser==0.15
docutils==0.19
docx==0.2.4
docx2txt==0.8
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==0.8.3
fastjsonschema==2.15.3
Flask==2.2.2
fonttools==4.34.4
fqdn==1.5.1
functions==0.0.0
future==0.18.3
hyperopt==0.2.7
idna==3.4
imagesize==1.4.1
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonpickle==3.0.1
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.22.1
kiwisolver==1.4.4
lazy_loader==0.1
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
md-mermaid==0.1.1
mistune==0.8.4
mypy-extensions==1.0.0
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.4
notebook==6.4.12
notebook_shim==0.2.3
numpy==1.23.0
openpyxl==3.0.10
overrides==7.3.1
p2j==1.3.2
packaging==21.3
pandas==1.4.3
pandas-flavor==0.5.0
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.5.1
playsound==1.2.2
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.3
psutil==5.9.1
pure-eval==0.2.2
py4j==0.10.9.7
pycparser==2.21
pydot==1.4.2
Pygments==2.14.0
pygraphviz @ file:///D:/bld/pygraphviz_1654467390271/work
pyodide-py==0.23.2
Pyomo==6.6.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-docx==0.8.11
python-json-logger==2.0.7
pytz==2022.1
pyvis==0.3.2
pywin32==304
pywinpty==2.0.6
PyYAML==6.0
pyzmq==25.1.0
rdkit-pypi==2022.3.3
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-learn==1.2.2
scipy==1.8.1
seaborn==0.13.2
selfies==2.1.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==6.1.3
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.3.0
tensorboardX==2.6
terminado==0.15.0
threadpoolctl==3.1.0
timeout-decorator==0.5.0
tinycss2==1.1.1
tomli==2.0.1
torch==1.13.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typed-argument-parser==1.8.0
typing-inspect==0.8.0
typing_extensions==4.5.0
uri-template==1.2.0
urllib3==1.26.15
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.2
Werkzeug==2.2.2
xarray==2023.2.0
